requirements.txt
ailment==9.2.76
angr==9.2.76
archinfo==9.2.76
asttokens==2.4.1
backports.strenum==1.2.8
bcrypt==4.0.1
bitarray==2.8.3
bitstring==4.1.2
cachetools==5.3.2
capstone==5.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
claripy==9.2.76
cle==9.2.76
colored-traceback==0.3.0
CppHeaderParser==2.7.4
cryptography==41.0.5
decorator==5.1.1
dpkt==1.9.8
exceptiongroup==1.1.3
executing==2.0.1
future==0.18.3
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
idna==3.4
intervaltree==3.1.0
ipython==8.17.2
itanium-demangler==1.1
jedi==0.19.1
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
mulpyplexer==0.9
nampa==0.1.1
networkx==3.2.1
packaging==23.2
paramiko==3.3.1
parso==0.8.3
pefile==2023.2.7
pexpect==4.8.0
plumbum==1.8.2
ply==3.11
prompt-toolkit==3.0.41
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pwntools==4.11.0
py==1.11.0
pycparser==2.21
pyelftools==0.30
Pygments==2.16.1
PyNaCl==1.5.0
pyserial==3.5
PySMT==0.9.5
PySocks==1.7.1
python-dateutil==2.8.2
pyvex==9.2.76
r2pipe==1.8.2
requests==2.31.0
rich==13.6.0
ROPGadget==7.4
rpyc==5.3.1
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.12
traitlets==5.13.0
unicorn==2.0.1.post1
urllib3==2.1.0
wcwidth==0.2.10
welpwn==0.97
z3-solver==4.10.2.0
zope.event==5.0
zope.interface==6.1
